feat!: SharedFiles now returns PublicFileMetadata (#24)
Added ShareMetadata argument to orchestrator endpoints. Added two maps to store those metadata. One is indexed by (UserCanister, FileId) -> FileMetadata, and the other is a reference counter to know how many users are sharing that file. Actually we can USE this map to also know how many users have access to that file. When the map is empty, we can finally remove the metadata to preserve storage

BREAKING CHANGE: Changed the api

Co-authored-by: panch8 <98569130+panch8@users.noreply.github.com>