feat(user-canister)!: Files are now indexed ALSO with their Path (#30)
A Path has been introduced to allow folders and a more structural file system for the UI. There is not (at least for now) FS tree, but they work as in S3, so the File Path it's the file name and from that the folders are derived. This required some API change to provide the File Path instead of the File name in the candid interface. The file path has been implemented as a new Type which wraps a String, but with some custom validation. For instance if a Path is being deserialized and it's either empty or it won't start with /, the deserialization will fail.

BREAKING CHANGE: changed request and response for request_file and upload_file_atomic